Update KDE Framewokrs to 5.69.0
April 11, 2020. KDE today announces the release of KDE Frameworks 5.69.0.

KDE Frameworks are over 70 addon libraries to Qt which provide a wide variety
of commonly needed functionality in mature, peer reviewed and well tested
libraries with friendly licensing terms. For an introduction see the KDE
Frameworks web page.

This release is part of a series of planned monthly releases making
improvements available to developers in a quick and predictable manner.

Announcement:
	https://kde.org/announcements/kde-frameworks-5.69.0.php