graphics/gdal: Add more default options
GDAL is an essential component for GIS software as graphics/qgis*, math/saga,
databases/postgis*, databases/grass*.
Unfortunately too many options are not enabled by default, causing the user
of theses softwares to rebuild graphics/gdal and its bindings (especially
graphics/py-gdal) for basic use.

With this commit, the vast majority of user needs will be covered, leaving
the build of graphics/gdal with the less relevant and less used options to
the users who really need it.

As gdal now requires postgresql, the package will install the default
version (postgresql12), however databases/postgis23 cannot be built with
postgresql>10 so it is marked as BROKEN -- however this is not a problem
since it is largely deprecated and will be removed soon.

PR:		224004
Approved by:	Maintainer timeout