net/unison232: isolate self-test from pre-existing $HOME
On the "exp - 121amd64-default-build-as-use" builder, $HOME appears
to be /nonexistent, and this jeopardizes the self-test.

Feed the selftest mktemp -d as its HOME so it doesn't read from,
or pollute, the actual building user's home directory.

Thanks to: pkg-fallout, kevans, rhtoad.

No PORTREVISION bump as this is purely a build-time fix.

MFH:		2020Q4 (blanket, one-line tested build robustness fix)