Switch x11-fonts/xorg-fonts-* to USES=metaport
Also, remove bugus CONFLICTS.

I believe that the FONTDIR could be renamed to _FONTDIR and the RUN_DEPENDS
sorted, but let's do it another time to keep the diff small. Also,
prefixing non-standard variables with "_" is not officially documented yet
in the Porter's Handbook (it's only the convention suggested by portclippy,
which itself is not mentioned there yet).

Reviewed by:	zeising (earlier version)
Approved by:	portmgr blanket
Differential Revision:	https://reviews.freebsd.org/D26989