Update to FreeBSD-CURRENT base SVN revision r367207
This update improves the internal pre-processor and makes it report data
file issues with filename and line number of the failed commend.

Several tests have been added to detect unmatched condition directives
(e.g. #endif or #else without prior #ifdef or #ifndef). An #undef command
has been added since it occurs in one of the calendar files tradionally
distributed with FreeBSD (and now in the deskutils/calendar-data port).

This update adds a default definition of _PATH_LOCALBASE that is required
to build the port on systems that do not define it in paths.h.