devel/llvm15: add a lite FLAVOR
The lite flavor is for when you just need a FreeBSD compiler (e.g., in
CI).  By default it includes CLANG and LLD as well as backends required
for FreeBSD and FreeBSD ports (BE_NATIVE, BE_AMDGPU, and BE_WASM).

The lite install is ~350GiB on amd64 vs ~1.7GiB for default and it has
only a single dependency (zstandard).  (The goal was no dependencies,
but libzstd is a private lib in base.)

Bump PORT_REVISION due to removal of clang libxml2 dependency.

Sponsored by:	DARPA