Added CONFLICTS with drivers and Fixed a mistake.
In x11/plasma6-plasma-desktop/Makefile:
Added CONFLICTS with the X.Org variants of the input drivers to the
'xlibre' flavor, and with XLibre variants of them to the 'xorg' flavor.

This is to ensure proper build time behaviour.

In x11/plasma6-plasma/Makefile:
RUN_DEPENDS was mistakenly assigned with the value of BUILD_DEPENDS.
Fixed it.